# specs/001-interview-evidence-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Interview Evidence Platform API
  version: 1.0.0
  description: |
    Contract baseline for the Korean-language pilot. All company routes enforce the company_id
    from the authenticated principal. Applicant routes enforce the invitation-scoped session
    cookie; neither API accepts company_id from an untrusted request body.
servers:
  - url: /v1
...
paths:
...
  /applicant/access/exchange:
    post:
      operationId: exchangeApplicantInvitationToken
      x-owner-lane: A
      tags: [Applicant Access]
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicantTokenExchange"
      responses:
        "204":
          description: |
            Scoped, expiring HttpOnly session cookie set. Raw invitation token is not returned.
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /applicant/access/revoke:
    post:
      operationId: revokeApplicantSession
      x-owner-lane: A
      tags: [Applicant Access]
      security:
        - ApplicantSession: []
      responses:
        "204":
          description: Current applicant session revoked and cookie cleared
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /applicant/identity-verifications:
    post:
      operationId: verifyApplicantIdentity
      x-owner-lane: A
      tags: [Applicant Access]
      security:
        - ApplicantSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicantIdentityVerification"
      responses:
        "200":
          description: Verified applicant access state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantAccessState"
        default:
          $ref: "#/components/responses/Error"
  /applicant/consents:
    get:
      operationId: getApplicantConsentPolicy
      x-owner-lane: A
      tags: [Applicant Access]
      security:
        - ApplicantSession: []
      responses:
        "200":
          description: Server-versioned policy content displayed before consent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentPolicyView"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: recordApplicantConsent
      x-owner-lane: A
      tags: [Applicant Access]
      security:
        - ApplicantSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentCreate"
      responses:
        "201":
          description: Immutable consent record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentView"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    CompanyBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApplicantSession:
      type: apiKey
      in: cookie
      name: iep_applicant_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  responses:
    Error:
      description: Structured error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id, retryable]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
          format: uuid
        retryable:
          type: boolean
        current_version:
          type: integer
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
...
    ApplicantTokenExchange:
      type: object
      additionalProperties: false
      required: [invitation_token]
      properties:
        invitation_token: {type: string, minLength: 32, maxLength: 4096}
    ApplicantIdentityVerification:
      type: object
      additionalProperties: false
      required: [display_name, verification_value]
      properties:
        display_name: {type: string, minLength: 1, maxLength: 200}
        verification_value: {type: string, minLength: 1, maxLength: 500}
    ApplicantAccessState:
      type: object
      additionalProperties: false
      required: [invitation_id, state, expires_at, required_actions]
      properties:
        invitation_id: {type: string, format: uuid}
        state: {type: string}
        expires_at: {type: string, format: date-time}
        required_actions:
          type: array
          items: {type: string}
    ConsentCreate:
      type: object
      additionalProperties: false
      required: [policy_version, accepted_purposes, consent_content_digest]
      properties:
        policy_version: {type: string}
        accepted_purposes:
          type: array
          minItems: 1
          items:
            type: string
            enum: [document_analysis, recording, ai_assessment]
        consent_content_digest: {type: string, pattern: "^[a-f0-9]{64}$"}
    ConsentPolicyView:
      type: object
      additionalProperties: false
      required:
        [policy_version, ai_role, recording_notice, processing_purposes,
         retention_days, deletion_method, required_purposes, content_digest]
      properties:
        policy_version: {type: string, minLength: 1}
        ai_role: {type: string, minLength: 1}
        recording_notice: {type: string, minLength: 1}
        processing_purposes:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [purpose, title, description]
            properties:
              purpose:
                type: string
                enum: [document_analysis, recording, ai_assessment]
              title: {type: string, minLength: 1}
              description: {type: string, minLength: 1}
        retention_days: {type: integer, minimum: 1}
        deletion_method: {type: string, minLength: 1}
        required_purposes:
          type: array
          minItems: 1
          items:
            type: string
            enum: [document_analysis, recording, ai_assessment]
        content_digest: {type: string, pattern: "^[a-f0-9]{64}$"}
    ConsentView:
      type: object
      additionalProperties: false
      required:
        [consent_record_id, policy_version, accepted_purposes, retention_days, accepted_at]
      properties:
        consent_record_id: {type: string, format: uuid}
        policy_version: {type: string}
        accepted_purposes:
          type: array
          items: {type: string}
        retention_days: {type: integer, minimum: 1}
        accepted_at: {type: string, format: date-time}
...